Add to ci rules, add failures to test ci
